Fix V8 NativePointer construction from large numbers
Where instead of clipping we would end up with zero on 32-bit systems.

This issue was caught by our GumJS NativePointer test-suite, but went
unnoticed due to earlier tests failing on our CI.

+++ bindings/gumjs/gumv8value.cpp
@@ -1311,7 +1311,7 @@ _gum_v8_native_pointer_parse (Local<Value> value,
       return TRUE;
     }
 
-    *ptr = GSIZE_TO_POINTER (number);
+    *ptr = GSIZE_TO_POINTER ((guint64) number);
     return TRUE;
   }
   else
